test: move common code
