+++ apps/ledger-live-desktop/tests/page/speculos.page.ts
@@ -1,31 +1,19 @@
 import { AppPage } from "tests/page/abstractClasses";
 import { step } from "tests/misc/reporters/step";
 import {
-  pressBoth,
-  pressUntilTextFound,
-  containsSubstringInEvent,
   activateLedgerSync,
   expectValidAddressDevice,
+  signSendTransaction,
+  signSendNFTTransaction,
+  signDelegationTransaction,
+  verifyAmountsAndAcceptSwap,
+  verifyAmountsAndRejectSwap,
 } from "@ledgerhq/live-common/e2e/speculos";
 import { Account } from "@ledgerhq/live-common/e2e/enum/Account";
-import { expect } from "@playwright/test";
-import { NFTTransaction, Transaction } from "tests/models/Transaction";
-import { Delegate } from "tests/models/Delegate";
-import { DeviceLabels } from "@ledgerhq/live-common/e2e/enum/DeviceLabels";
-import { Currency } from "@ledgerhq/live-common/e2e/enum/Currency";
-import { Swap } from "tests/models/Swap";
-import { extractNumberFromString } from "tests/utils/textParserUtils";
-import { sendBTCBasedCoin } from "tests/families/bitcoin";
-import { sendEVM, sendEvmNFT } from "tests/families/evm";
-import { sendPolkadot } from "tests/families/polkadot";
-import { sendAlgorand } from "tests/families/algorand";
-import { sendTron } from "tests/families/tron";
-import { sendStellar } from "tests/families/stellar";
-import { sendCardano } from "tests/families/cardano";
-import { sendXRP } from "tests/families/xrp";
-import { delegateNear } from "tests/families/near";
-import { delegateCosmos, sendCosmos } from "tests/families/cosmos";
-import { delegateSolana, sendSolana } from "tests/families/solana";
+import { NFTTransaction, Transaction } from "@ledgerhq/live-common/e2e/models/Transaction";
+import { Delegate } from "@ledgerhq/live-common/e2e/models/Delegate";
+
+import { Swap } from "@ledgerhq/live-common/e2e/models/Swap";
 export class SpeculosPage extends AppPage {
   @step("Verify receive address correctness on device")
   async expectValidAddressDevice(account: Account, addressDisplayed: string) {
@@ -39,113 +27,26 @@ export class SpeculosPage extends AppPage {
 
   @step("Sign Send NFT Transaction")
   async signSendNFTTransaction(tx: NFTTransaction) {
-    const currencyName = tx.accountToDebit.currency;
-    switch (currencyName) {
-      case Currency.ETH:
-        await sendEvmNFT(tx);
-        break;
-      default:
-        throw new Error(`Unsupported currency: ${currencyName}`);
-    }
+    await signSendNFTTransaction(tx);
   }
 
   @step("Sign Send Transaction")
   async signSendTransaction(tx: Transaction) {
-    const currencyName = tx.accountToDebit.currency;
-    switch (currencyName) {
-      case Currency.sepETH:
-      case Currency.POL:
-        await sendEVM(tx);
-        break;
-      case Currency.DOGE:
-      case Currency.BCH:
-        await sendBTCBasedCoin(tx);
-        break;
-      case Currency.DOT:
-        await sendPolkadot(tx);
-        break;
-      case Currency.ALGO:
-        await sendAlgorand(tx);
-        break;
-      case Currency.SOL:
-        await sendSolana(tx);
-        break;
-      case Currency.TRX:
-        await sendTron(tx);
-        break;
-      case Currency.XLM:
-        await sendStellar(tx);
-        break;
-      case Currency.ATOM:
-        await sendCosmos(tx);
-        break;
-      case Currency.ADA:
-        await sendCardano(tx);
-        break;
-      case Currency.XRP:
-        await sendXRP(tx);
-        break;
-      default:
-        throw new Error(`Unsupported currency: ${currencyName}`);
-    }
+    await signSendTransaction(tx);
   }
 
   @step("Sign Delegation Transaction")
   async signDelegationTransaction(delegatingAccount: Delegate) {
-    const currencyName = delegatingAccount.account.currency.name;
-    switch (currencyName) {
-      case Account.SOL_1.currency.name:
-        await delegateSolana();
-        break;
-      case Account.NEAR_1.currency.name:
-        await delegateNear(delegatingAccount);
-        break;
-      case Account.ATOM_1.currency.name:
-        await delegateCosmos(delegatingAccount);
-        break;
-      default:
-        throw new Error(`Unsupported currency: ${currencyName}`);
-    }
+    await signDelegationTransaction(delegatingAccount);
   }
 
   @step("Verify amounts and accept swap")
   async verifyAmountsAndAcceptSwap(swap: Swap) {
-    const events = await pressUntilTextFound(DeviceLabels.ACCEPT);
-    await this.verifySwapData(swap, events);
-    await pressBoth();
+    await verifyAmountsAndAcceptSwap(swap);
   }
 
   @step("Verify amounts and reject swap")
   async verifyAmountsAndRejectSwap(swap: Swap) {
-    const events = await pressUntilTextFound(DeviceLabels.REJECT);
-    await this.verifySwapData(swap, events);
-    await pressBoth();
-  }
-
-  async verifySwapData(swap: Swap, events: string[]) {
-    const sendAmountScreen = containsSubstringInEvent(swap.amount, events);
-    expect(sendAmountScreen).toBeTruthy();
-    this.verifySwapGetAmountScreen(swap, events);
-    this.verifySwapFeesAmountScreen(swap, events);
-  }
-
-  verifySwapGetAmountScreen(swap: Swap, events: string[]) {
-    const parsedAmountToReceive = extractNumberFromString(swap.amountToReceive);
-    swap.amountToReceive =
-      parsedAmountToReceive.length < 19
-        ? parsedAmountToReceive
-        : parsedAmountToReceive.substring(0, 18);
-
-    const receivedGetAmount = containsSubstringInEvent(`${swap.amountToReceive}`, events);
-    expect(receivedGetAmount).toBeTruthy();
-  }
-
-  verifySwapFeesAmountScreen(swap: Swap, events: string[]) {
-    const parsedFeesAmount = extractNumberFromString(swap.feesAmount);
-    swap.feesAmount =
-      parsedFeesAmount.length < 19 ? parsedFeesAmount : parsedFeesAmount.substring(0, 18);
-
-    const receivedFeesAmount = containsSubstringInEvent(swap.feesAmount, events);
-    expect(receivedFeesAmount).toBeTruthy();
+    await verifyAmountsAndRejectSwap(swap);
   }
 }
