+++ apps/ledger-live-desktop/tests/specs/speculos/send.tx.spec.ts
@@ -2,7 +2,7 @@ import { test } from "../../fixtures/common";
 import { Account } from "@ledgerhq/live-common/e2e/enum/Account";
 import { Fee } from "@ledgerhq/live-common/e2e/enum/Fee";
 import { OperationType } from "@ledgerhq/live-common/e2e/enum/OperationType";
-import { Transaction } from "../../models/Transaction";
+import { Transaction } from "@ledgerhq/live-common/e2e/models/Transaction";
 import { addTmsLink, addBugLink } from "tests/utils/allureUtils";
 import { getDescription } from "../../utils/customJsonReporter";
 import { CLI } from "tests/utils/cliUtils";
